safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/inline-style-blocked-expected.txt

CONSOLE MESSAGE: line 5: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'none'".

PASS
